PR Changes- After automation of generating the tokens
