+++ src/test/java/iudx/resource/server/apiserver/integrationtests/temporalSearchAPIs/GetAfterTemporalEntitiesIT.java
@@ -1,4 +1,4 @@
-package iudx.resource.server.apiserver.integrationtests.TemporalSearchAPIs;
+package iudx.resource.server.apiserver.integrationtests.temporalSearchAPIs;
 
 import iudx.resource.server.apiserver.integrationtests.RestAssuredConfiguration;
 import org.junit.jupiter.api.DisplayName;
@@ -6,21 +6,21 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 import static io.restassured.RestAssured.given;
-import static iudx.resource.server.authenticator.JwtTokenHelper.openResourceToken;
+
+import static iudx.resource.server.authenticator.TokensForITs.openResourceToken;
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.Matchers.equalTo;
 
 @ExtendWith(RestAssuredConfiguration.class)
-public class getBetweenTemporalEntitiesIT {
+public class GetAfterTemporalEntitiesIT {
     String temporalId="b58da193-23d9-43eb-b98a-a103d4b6103c";
     @Test
-    @DisplayName("200 (success) temporal (between)")
+    @DisplayName("200 (success) temporal (after)")
     public void getTemporalEntityTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerel", "during")
-                .queryParam("time", "2020-10-18T14:20:00Z")
-                .queryParam("endtime", "2020-10-19T14:20:00Z")
+                .queryParam("timerel", "after")
+                .queryParam("time", "2020-10-18T14:20:01Z")
                 .header("Content-Type", "application/json")
                 .header("token", openResourceToken)
                 .when()
@@ -33,32 +33,12 @@ public class getBetweenTemporalEntitiesIT {
                 .body("results[0].id", notNullValue());
     }
     @Test
-    @DisplayName("200 (success) temporal (between) with optional encryption")
-    public void getTemporalEntityWithOptionalEncryptionTest(){
-        given()
-                .queryParam("id",temporalId)
-                .queryParam("timerel", "during")
-                .queryParam("time", "2020-10-18T14:20:00Z")
-                .queryParam("endtime", "2020-10-19T14:20:00Z")
-                .header("Content-Type", "application/json")
-                .header("token", openResourceToken)
-                .when()
-                .get("/temporal/entities")
-                .then()
-                .statusCode(200)
-                .log().body()
-                .body("title", equalTo("Success"))
-                .body("type", equalTo("urn:dx:rs:success"))
-                .body("results[0].id", notNullValue());
-    }
-    @Test
-    @DisplayName("204 (Empty Response) temporal (between)")
+    @DisplayName("204 (Empty Response) temporal (after)")
     public void getTemporalEntityWithEmptyResponseTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerel", "during")
-                .queryParam("time", "2020-01-18T14:20:00Z")
-                .queryParam("endtime", "2020-01-19T14:20:00Z")
+                .queryParam("timerel", "after")
+                .queryParam("time", "2020-01-19T14:20:00Z")
                 .header("token", openResourceToken)
                 .when()
                 .get("/temporal/entities")
@@ -67,13 +47,12 @@ public class getBetweenTemporalEntitiesIT {
                 .log().body();
     }
     @Test
-    @DisplayName("400 (invalid params) temporal (between)")
+    @DisplayName("400 (Invalid params) temporal (after)")
     public void getTemporalEntityWithInvalidParamsTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerelation", "during")
-                .queryParam("time", "2020-09-18T14:20:00Z")
-                .queryParam("endtime", "2020-09-19T14:20:00Z")
+                .queryParam("timerelation", "after")
+                .queryParam("timea", "2020-06-01T14:20:01Z")
                 .header("Content-Type", "application/json")
                 .header("token", openResourceToken)
                 .when()
@@ -85,13 +64,12 @@ public class getBetweenTemporalEntitiesIT {
                 .body("type", equalTo("urn:dx:rs:invalidTemporalRelationParam"));
     }
     @Test
-    @DisplayName("400 (invalid date format) temporal (between)")
+    @DisplayName("400 (invalid date format) temporal (after)")
     public void getTemporalEntityWithInvalidDateFormatTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerel", "during")
-                .queryParam("time", "2020-09-18X14:20:00Z")
-                .queryParam("endtime", "2020-09-19X14:20:00Z")
+                .queryParam("timerel", "after")
+                .queryParam("time", "2020-06-01X14:20:01Z")
                 .header("Content-Type", "application/json")
                 .header("token", openResourceToken)
                 .when()
@@ -103,14 +81,13 @@ public class getBetweenTemporalEntitiesIT {
                 .body("type", equalTo("urn:dx:rs:invalidAttributeValue"));
     }
     @Test
-    @DisplayName("404 (not found) temporal (between)")
+    @DisplayName("404 (not found) temporal (after)")
     public void TemporalEntityNotFoundTest(){
-        String nonExistingTemporalId="b58da193-23d9-43eb-b98a-a103d4b61030";
+        String nonExistingTemporalId="b58da193-23d9-43eb-b98a-a103d4b6102c";
         given()
                 .queryParam("id",nonExistingTemporalId)
-                .queryParam("timerel", "during")
-                .queryParam("time", "2020-09-18T14:20:00Z")
-                .queryParam("endtime", "2020-09-19T14:20:00Z")
+                .queryParam("timerel", "after")
+                .queryParam("time", "2020-06-01T14:20:01Z")
                 .header("Content-Type", "application/json")
                 .header("token", openResourceToken)
                 .when()
@@ -122,13 +99,12 @@ public class getBetweenTemporalEntitiesIT {
                 .body("type", equalTo("urn:dx:rs:resourceNotFound"));
     }
     @Test
-    @DisplayName("401(invalid credentials) temporal (between)")
+    @DisplayName("401(invalid credentials) temporal (after)")
     public void getTemporalEntityWithInvalidCredentialsTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerel", "during")
-                .queryParam("time", "2020-10-18T14:20:00Z")
-                .queryParam("endtime", "2020-10-19T14:20:00Z")
+                .queryParam("timerel", "after")
+                .queryParam("time", "2020-06-01T14:20:01Z")
                 .header("Content-Type", "application/json")
                 .header("token", "abc")
                 .when()
@@ -139,6 +115,5 @@ public class getBetweenTemporalEntitiesIT {
                 .body("title", equalTo("Not Authorized"))
                 .body("type", equalTo("urn:dx:rs:invalidAuthorizationToken"));
     }
-
 
 }
