+++ src/test/java/iudx/resource/server/apiserver/integrationtests/temporalSearchAPIs/GetBeforeTemporalEntitiesIT.java
@@ -1,4 +1,4 @@
-package iudx.resource.server.apiserver.integrationtests.TemporalSearchAPIs;
+package iudx.resource.server.apiserver.integrationtests.temporalSearchAPIs;
 
 import iudx.resource.server.apiserver.integrationtests.RestAssuredConfiguration;
 import org.junit.jupiter.api.DisplayName;
@@ -6,20 +6,21 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 import static io.restassured.RestAssured.given;
-import static iudx.resource.server.authenticator.JwtTokenHelper.openResourceToken;
+
+import static iudx.resource.server.authenticator.TokensForITs.openResourceToken;
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.Matchers.equalTo;
 
 @ExtendWith(RestAssuredConfiguration.class)
-public class getAfterTemporalEntitiesIT {
+public class GetBeforeTemporalEntitiesIT {
     String temporalId="b58da193-23d9-43eb-b98a-a103d4b6103c";
     @Test
-    @DisplayName("200 (success) temporal (after)")
+    @DisplayName("200 (success) temporal (before)")
     public void getTemporalEntityTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerel", "after")
-                .queryParam("time", "2020-10-18T14:20:01Z")
+                .queryParam("timerel", "before")
+                .queryParam("time", "2020-10-19T12:00:00Z")
                 .header("Content-Type", "application/json")
                 .header("token", openResourceToken)
                 .when()
@@ -32,11 +33,11 @@ public class getAfterTemporalEntitiesIT {
                 .body("results[0].id", notNullValue());
     }
     @Test
-    @DisplayName("204 (Empty Response) temporal (after)")
+    @DisplayName("204 (Empty Response) temporal (before)")
     public void getTemporalEntityWithEmptyResponseTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerel", "after")
+                .queryParam("timerel", "before")
                 .queryParam("time", "2020-01-19T14:20:00Z")
                 .header("token", openResourceToken)
                 .when()
@@ -46,29 +47,12 @@ public class getAfterTemporalEntitiesIT {
                 .log().body();
     }
     @Test
-    @DisplayName("400 (Invalid params) temporal (after)")
+    @DisplayName("400 (Invalid params) temporal (before)")
     public void getTemporalEntityWithInvalidParamsTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerelation", "after")
-                .queryParam("timea", "2020-06-01T14:20:01Z")
-                .header("Content-Type", "application/json")
-                .header("token", openResourceToken)
-                .when()
-                .get("/temporal/entities")
-                .then()
-                .statusCode(400)
-                .log().body()
-                .body("title", equalTo("Bad Request"))
-                .body("type", equalTo("urn:dx:rs:invalidTemporalRelationParam"));
-    }
-    @Test
-    @DisplayName("400 (invalid date format) temporal (after)")
-    public void getTemporalEntityWithInvalidDateFormatTest(){
-        given()
-                .queryParam("id",temporalId)
-                .queryParam("timerel", "after")
-                .queryParam("time", "2020-06-01X14:20:01Z")
+                .queryParam("timerel", "before")
+                .queryParam("timea", "2020-09-19T14:20:00Z")
                 .header("Content-Type", "application/json")
                 .header("token", openResourceToken)
                 .when()
@@ -80,13 +64,30 @@ public class getAfterTemporalEntitiesIT {
                 .body("type", equalTo("urn:dx:rs:invalidAttributeValue"));
     }
     @Test
-    @DisplayName("404 (not found) temporal (after)")
-    public void TemporalEntityNotFoundTest(){
+    @DisplayName("400 (invalid date format) temporal (before)")
+    public void getTemporalEntityWithInvalidDateFormatTest(){
+        given()
+                .queryParam("id",temporalId)
+                .queryParam("timerel", "before")
+                .queryParam("time", "2020-09-19X14:20:00Z")
+                .header("Content-Type", "application/json")
+                .header("token", openResourceToken)
+                .when()
+                .get("/temporal/entities")
+                .then()
+                .statusCode(400)
+                .log().body()
+                .body("title", equalTo("Bad Request"))
+                .body("type", equalTo("urn:dx:rs:invalidAttributeValue"));
+    }
+    @Test
+    @DisplayName("404 (not found) temporal (before)")
+    public void temporalEntityNotFoundTest(){
         String nonExistingTemporalId="b58da193-23d9-43eb-b98a-a103d4b6102c";
         given()
                 .queryParam("id",nonExistingTemporalId)
-                .queryParam("timerel", "after")
-                .queryParam("time", "2020-06-01T14:20:01Z")
+                .queryParam("timerel", "before")
+                .queryParam("time", "2020-09-19T14:20:00Z")
                 .header("Content-Type", "application/json")
                 .header("token", openResourceToken)
                 .when()
@@ -98,12 +99,12 @@ public class getAfterTemporalEntitiesIT {
                 .body("type", equalTo("urn:dx:rs:resourceNotFound"));
     }
     @Test
-    @DisplayName("401(invalid credentials) temporal (after)")
+    @DisplayName("401(invalid credentials) temporal (before)")
     public void getTemporalEntityWithInvalidCredentialsTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerel", "after")
-                .queryParam("time", "2020-06-01T14:20:01Z")
+                .queryParam("timerel", "before")
+                .queryParam("time", "2020-10-19T14:20:00Z")
                 .header("Content-Type", "application/json")
                 .header("token", "abc")
                 .when()
@@ -114,5 +115,4 @@ public class getAfterTemporalEntitiesIT {
                 .body("title", equalTo("Not Authorized"))
                 .body("type", equalTo("urn:dx:rs:invalidAuthorizationToken"));
     }
-
 }
