+++ src/test/java/iudx/resource/server/apiserver/integrationtests/temporalSearchAPIs/GetBetweenTemporalEntitiesIT.java
@@ -1,4 +1,4 @@
-package iudx.resource.server.apiserver.integrationtests.TemporalCountAPIs;
+package iudx.resource.server.apiserver.integrationtests.temporalSearchAPIs;
 
 import iudx.resource.server.apiserver.integrationtests.RestAssuredConfiguration;
 import org.junit.jupiter.api.DisplayName;
@@ -6,22 +6,22 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 
 import static io.restassured.RestAssured.given;
-import static iudx.resource.server.authenticator.JwtTokenHelper.openResourceToken;
+
+import static iudx.resource.server.authenticator.TokensForITs.openResourceToken;
 import static org.hamcrest.CoreMatchers.notNullValue;
 import static org.hamcrest.Matchers.equalTo;
-import static org.hamcrest.Matchers.is;
 
 @ExtendWith(RestAssuredConfiguration.class)
-public class getBeforeTemporalCountIT {
+public class GetBetweenTemporalEntitiesIT {
     String temporalId="b58da193-23d9-43eb-b98a-a103d4b6103c";
     @Test
-    @DisplayName("200 (success) temporal (before)")
-    public void getTemporalCountTest(){
+    @DisplayName("200 (success) temporal (between)")
+    public void getTemporalEntityTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerel", "before")
-                .queryParam("time", "2020-10-19T12:00:00Z")
-                .queryParam("options","count")
+                .queryParam("timerel", "during")
+                .queryParam("time", "2020-10-18T14:20:00Z")
+                .queryParam("endtime", "2020-10-19T14:20:00Z")
                 .header("Content-Type", "application/json")
                 .header("token", openResourceToken)
                 .when()
@@ -31,17 +31,35 @@ public class getBeforeTemporalCountIT {
                 .log().body()
                 .body("title", equalTo("Success"))
                 .body("type", equalTo("urn:dx:rs:success"))
-                .body("results[0]", notNullValue())
-                .body("results[0].totalHits", is(notNullValue()));
+                .body("results[0].id", notNullValue());
     }
     @Test
-    @DisplayName("204 (Empty Response) temporal (before)")
-    public void getTemporalCountWithEmptyResponseTest(){
+    @DisplayName("200 (success) temporal (between) with optional encryption")
+    public void getTemporalEntityWithOptionalEncryptionTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerel", "before")
-                .queryParam("time", "2020-01-19T12:00:00Z")
-                .queryParam("options","count")
+                .queryParam("timerel", "during")
+                .queryParam("time", "2020-10-18T14:20:00Z")
+                .queryParam("endtime", "2020-10-19T14:20:00Z")
+                .header("Content-Type", "application/json")
+                .header("token", openResourceToken)
+                .when()
+                .get("/temporal/entities")
+                .then()
+                .statusCode(200)
+                .log().body()
+                .body("title", equalTo("Success"))
+                .body("type", equalTo("urn:dx:rs:success"))
+                .body("results[0].id", notNullValue());
+    }
+    @Test
+    @DisplayName("204 (Empty Response) temporal (between)")
+    public void getTemporalEntityWithEmptyResponseTest(){
+        given()
+                .queryParam("id",temporalId)
+                .queryParam("timerel", "during")
+                .queryParam("time", "2020-01-18T14:20:00Z")
+                .queryParam("endtime", "2020-01-19T14:20:00Z")
                 .header("token", openResourceToken)
                 .when()
                 .get("/temporal/entities")
@@ -50,13 +68,31 @@ public class getBeforeTemporalCountIT {
                 .log().body();
     }
     @Test
-    @DisplayName("400 (Invalid params) temporal (before)")
-    public void getTemporalCountWithInvalidParamsTest(){
+    @DisplayName("400 (invalid params) temporal (between)")
+    public void getTemporalEntityWithInvalidParamsTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerel", "before")
-                .queryParam("timeattribute", "2020-09-19T12:00:00Z")
-                .queryParam("options","count")
+                .queryParam("timerelation", "during")
+                .queryParam("time", "2020-09-18T14:20:00Z")
+                .queryParam("endtime", "2020-09-19T14:20:00Z")
+                .header("Content-Type", "application/json")
+                .header("token", openResourceToken)
+                .when()
+                .get("/temporal/entities")
+                .then()
+                .statusCode(400)
+                .log().body()
+                .body("title", equalTo("Bad Request"))
+                .body("type", equalTo("urn:dx:rs:invalidTemporalRelationParam"));
+    }
+    @Test
+    @DisplayName("400 (invalid date format) temporal (between)")
+    public void getTemporalEntityWithInvalidDateFormatTest(){
+        given()
+                .queryParam("id",temporalId)
+                .queryParam("timerel", "during")
+                .queryParam("time", "2020-09-18X14:20:00Z")
+                .queryParam("endtime", "2020-09-19X14:20:00Z")
                 .header("Content-Type", "application/json")
                 .header("token", openResourceToken)
                 .when()
@@ -68,32 +104,14 @@ public class getBeforeTemporalCountIT {
                 .body("type", equalTo("urn:dx:rs:invalidAttributeValue"));
     }
     @Test
-    @DisplayName("400 (invalid date format) temporal (before)")
-    public void getTemporalCountWithInvalidDateFormatTest(){
-        given()
-                .queryParam("id",temporalId)
-                .queryParam("timerel", "before")
-                .queryParam("time", "2020-09-19XX12:00:00Z")
-                .queryParam("options","count")
-                .header("Content-Type", "application/json")
-                .header("token", openResourceToken)
-                .when()
-                .get("/temporal/entities")
-                .then()
-                .statusCode(400)
-                .log().body()
-                .body("title", equalTo("Bad Request"))
-                .body("type", equalTo("urn:dx:rs:invalidAttributeValue"));
-    }
-    @Test
-    @DisplayName("404 (not found) temporal (before)")
-    public void temporalCountNotFoundTest(){
-        String nonExistingTemporalId="b58da193-23d9-43eb-b98a-a103d4b6107c";
+    @DisplayName("404 (not found) temporal (between)")
+    public void TemporalEntityNotFoundTest(){
+        String nonExistingTemporalId="b58da193-23d9-43eb-b98a-a103d4b61030";
         given()
                 .queryParam("id",nonExistingTemporalId)
-                .queryParam("timerel", "before")
-                .queryParam("time", "2020-09-19T12:00:00Z")
-                .queryParam("options","count")
+                .queryParam("timerel", "during")
+                .queryParam("time", "2020-09-18T14:20:00Z")
+                .queryParam("endtime", "2020-09-19T14:20:00Z")
                 .header("Content-Type", "application/json")
                 .header("token", openResourceToken)
                 .when()
@@ -105,13 +123,13 @@ public class getBeforeTemporalCountIT {
                 .body("type", equalTo("urn:dx:rs:resourceNotFound"));
     }
     @Test
-    @DisplayName("401(invalid credentials) temporal (before)")
-    public void getTemporalCountWithInvalidCredentialsTest(){
+    @DisplayName("401(invalid credentials) temporal (between)")
+    public void getTemporalEntityWithInvalidCredentialsTest(){
         given()
                 .queryParam("id",temporalId)
-                .queryParam("timerel", "before")
-                .queryParam("time", "2020-09-19T12:00:00Z")
-                .queryParam("options","count")
+                .queryParam("timerel", "during")
+                .queryParam("time", "2020-10-18T14:20:00Z")
+                .queryParam("endtime", "2020-10-19T14:20:00Z")
                 .header("Content-Type", "application/json")
                 .header("token", "abc")
                 .when()
@@ -122,4 +140,6 @@ public class getBeforeTemporalCountIT {
                 .body("title", equalTo("Not Authorized"))
                 .body("type", equalTo("urn:dx:rs:invalidAuthorizationToken"));
     }
+
+
 }
